requirements.txt
openai
transformers
torch
Pillow
python-dotenv
requests
pytest
